+++ config/expert/Skylines/code/CGCode.cpp
@@ -1,0 +1,23 @@
+#include <iostream>
+#include <string>
+#include <vector>
+#include <algorithm>
+
+using namespace std;
+
+int main()
+{
+    int n;
+    cin >> n; cin.ignore();
+    for (int i = 0; i < n; i++) {
+        int h;
+        int x_1;
+        int x_2;
+        cin >> h >> x_1 >> x_2; cin.ignore();
+    }
+
+    // Write an answer using cout. DON'T FORGET THE "<< endl"
+    // To debug: cerr << "Debug messages..." << endl;
+
+    cout << "linecount" << endl;
+}