+++ config/expert/Skylines/code/CGCode.java
@@ -1,0 +1,21 @@
+import java.util.*;
+import java.io.*;
+import java.math.*;
+
+class Solution {
+
+    public static void main(String args[]) {
+        Scanner in = new Scanner(System.in);
+        int n = in.nextInt();
+        for (int i = 0; i < n; i++) {
+            int h = in.nextInt();
+            int x1 = in.nextInt();
+            int x2 = in.nextInt();
+        }
+
+        // Write an answer using System.out.println()
+        // To debug: System.err.println("Debug messages...");
+
+        System.out.println("linecount");
+    }
+}