+++ config/expert/Skylines/code/CGCode.js
@@ -1,0 +1,12 @@
+const n = parseInt(readline());
+for (let i = 0; i < n; i++) {
+    var inputs = readline().split(' ');
+    const h = parseInt(inputs[0]);
+    const x1 = parseInt(inputs[1]);
+    const x2 = parseInt(inputs[2]);
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('linecount');