+++ config/expert/Skylines/code/CGCode.ts
@@ -1,0 +1,12 @@
+const n: number = parseInt(readline());
+for (let i = 0; i < n; i++) {
+    var inputs: string[] = readline().split(' ');
+    const h: number = parseInt(inputs[0]);
+    const x1: number = parseInt(inputs[1]);
+    const x2: number = parseInt(inputs[2]);
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('linecount');